omprog-sqlite: put syslog events in sqlite with db normalization
Instead of a flat list of syslog messages as is easily provided by the
omlibdbi module distributed with rsyslog, we split it up to a group of
tables:

  hosts
  facilities
  programs
  severities
  events

The columns of events refer to the other tables using foreign keys where
applicable.

Currently, there's a naive approach for trying to upsert new entries in
the tables. If we get for instance a severity id from rsyslog that we
haven't seen before, we will insert the new id and associate it with the
name supplied by rsyslog (i.e. the first one wins when it comes to
naming the {severity, facility} ids). For programs and hosts, the
situation is similar; but instead of being fed a id from rsyslog, we
invent it on our own by asking sqlite for it (using the AUTOINCREMENT
mechanism).

When working with a single sqlite3 database, you deal with the same
parallelism constraints that you would when dealing directly with
sqlite3. You should therefore probably not run two omprog-sqlite at the
same time (working against the same database). This means that you can't
have several rsyslog actions that all point to the same database. You
should instead either create rsyslog rulesets or use mutliple databases.

Currently, no attempts to clean the database of old entries are made.
I'll probably implement that at some point. (Probably using a cleanup
script run from cron. Easy does it.)

It's a quick hack. I wouldn't disable regular logging mechanisms when
using this, but maybe it turns out to work out fine. Let's wait and see!

I also implemented a simple "viewer", where you can query by
host/facility/program/severity (see rsyslog-db-viewer).

Finally, a simple rsyslog.conf fragment is included as an example.

Missing stuff:
 * Logging database errors (! this one is important, maybe?)
 * Query events by time ranges
 * DB cleanup
 * Colors?